Completed endpoint skeletons
There's still plenty that needs to be resolved. The queries haven't been tested as the database needs to be setup first. Some clarification needed for some queries regarding what the blueprint meant by certain values. These should be minor fixes, and we should be able to finish on time for demonstration.